+++ src/main/java/com/fastcampus/oop/logic/BubbleSort.java
@@ -1,4 +1,4 @@
-package com.fastcampus.oop;
+package com.fastcampus.oop.logic;
 
 import java.util.ArrayList;
 import java.util.List;
